Create documentation.md
This is a WIP documentation of the data packs functionality. Since the project was growing quite large I wanted to be able to reference certain things quickly and at a root level. This is the file for that. It's about 30% complete however and I will need to update as changes are made.